update demo pres 8/10

+++ demo_wait_file.py
@@ -20,15 +20,21 @@ if os.path.exists(file_path):
     os.remove(file_path)
 
 # Simuler la création du fichier après 5 secondes
-def create_file_later(file_path, delay, logger):
+def create_file_with_delay(file_path, delay, logger):
     time.sleep(delay)
     with open(file_path, 'w') as f:
         f.write("Fichier crée.")
     logger.info("Fichier crée par un autre processus")
 
-threading.Thread(target=ads.wait_for_file, args=(file_path, logger, 10)).start()
-threading.Thread(target=create_file_later, args=(file_path, 5, logger)).start()
 
-time.sleep(6)
-os.remove(file_path)
+threading.Thread(target=create_file_with_delay, args=(file_path, 5, logger)).start()
+logger.info("Début de la recherche.")
+if ads.wait_for_file(os.getcwd(), file_path, 10):
+    logger.info("Fichier bien trouvé.")
+else:
+    logger.error("Pas de fichier trouvé.")
+
+if os.path.exists(file_path):
+    os.remove(file_path)
 logger.info("Fin de la démonstration")
+
